+++ src/main/java/com/rockset/jdbc/RocksetUtils.java
@@ -26,7 +26,7 @@ class RocksetUtils {
         stype = java.sql.Types.NULL;
         break;
       case NUMBER: 
-        stype = java.sql.Types.BIGINT;
+        stype = java.sql.Types.DOUBLE;
         break;
       case OBJECT: 
       case POJO:
